refactor: remove redundant type casts

+++ scm-core/src/main/java/sonia/scm/util/IOUtil.java
@@ -229,7 +229,7 @@ public final class IOUtil
       while (byteCount > 0)
       {
         int max = (byteCount < bufferSize)
-          ? (int) byteCount
+          ? byteCount
           : bufferSize;
 
         len = in.read(buffer, 0, max);
